requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
bleach==3.1.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
jellyfish==0.7.2
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
lxml==4.3.4
MarkupSafe==1.1.1
mistune==0.8.4
mongoengine==0.18.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
pandocfilters==1.4.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pymongo==3.8.0
pyOpenSSL==19.0.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pyzmq==18.0.0
qtconsole==4.5.1
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.8
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
WTForms==2.2.1
